helper/analysisfixtest: New package to handle SuggestedFixes testing (#118)
The `go/analysis` framework can automatically adjust code via `Diagnostic.SuggestedFixes` to fix issues. This package contains a small testing framework for verifying the execution and file contents of those fixes.

The underlying `go/analysis` logic to invoke those fixes is unfortunately buried in internal packages so this package takes the worst case approach of creating a wrapped Analyzer in `singlecheck.Main()` and invoking the `-fix` flag in GOPATH mode to test it.