Ensure SchemaInfo.Schema.Default is non-nil if it is really set
In prior implementation, if the schema type is one of int/bool/string,
the `.Schema.Default` might not be set if the `Default` is not set to a
BasicLiteral(for int/string) or a bool Ident (for bool). It is a very
popular case that the default value of a string property is set to an
expression that is to convert the type from another string aliased type.