Remove terraform-plugin-sdk (v1) support (#305)
This dependency was marked end-of-life many years ago and it causes issues with attempting to support newer versions of terraform-plugin-sdk/v2 due to shared dependencies with breaking changes. This change removes the v1 dependency and empties out any analyzers with logic solely for that version. The analyzers will remain in place to reserve the analyzer name, but they will now never return reports.